:sparkles: Feat: Profile - MVVM 패턴 적용
