+++ lib/models/followuser_profile_model.dart
@@ -1,0 +1,6 @@
+class FollowuserProfileModel {
+  final String name;
+  final String description;
+
+  FollowuserProfileModel({required this.name, required this.description});
+}
