+++ lib/models/user_profile_model.dart
@@ -1,0 +1,11 @@
+class UserProfileModel {
+  final String name;
+  final String description;
+  final List<String> tags;
+
+  UserProfileModel(
+    this.name,
+    this.description,
+    this.tags,
+  );
+}
